AB2d-4457: Initial move of Job related classes to job module.  Got the Job tests working properly.
