+++ api/src/main/java/gov/cms/ab2d/api/controller/ErrorHandler.java
@@ -10,10 +10,10 @@ import gov.cms.ab2d.api.security.InvalidAuthHeaderException;
 import gov.cms.ab2d.api.security.MissingTokenException;
 import gov.cms.ab2d.common.service.InvalidClientInputException;
 import gov.cms.ab2d.common.service.InvalidContractException;
-import gov.cms.ab2d.common.service.InvalidJobAccessException;
-import gov.cms.ab2d.common.service.InvalidJobStateTransition;
+import gov.cms.ab2d.job.service.InvalidJobAccessException;
+import gov.cms.ab2d.job.service.InvalidJobStateTransition;
 import gov.cms.ab2d.common.service.InvalidPropertiesException;
-import gov.cms.ab2d.common.service.JobOutputMissingException;
+import gov.cms.ab2d.job.service.JobOutputMissingException;
 import gov.cms.ab2d.common.service.ResourceNotFoundException;
 import gov.cms.ab2d.eventlogger.Ab2dEnvironment;
 import gov.cms.ab2d.eventlogger.LogManager;
