+++ api/src/test/java/gov/cms/ab2d/api/remote/JobClientMock.java
@@ -5,9 +5,9 @@ import gov.cms.ab2d.common.dto.StartJobDTO;
 import gov.cms.ab2d.common.model.JobOutput;
 import gov.cms.ab2d.common.model.JobStatus;
 import gov.cms.ab2d.common.model.TooFrequentInvocations;
-import gov.cms.ab2d.common.service.InvalidJobStateTransition;
-import gov.cms.ab2d.common.service.JobOutputMissingException;
 import gov.cms.ab2d.common.service.ResourceNotFoundException;
+import gov.cms.ab2d.job.service.InvalidJobStateTransition;
+import gov.cms.ab2d.job.service.JobOutputMissingException;
 import gov.cms.ab2d.eventlogger.LogManager;
 import gov.cms.ab2d.eventlogger.events.JobStatusChangeEvent;
 import org.springframework.beans.factory.annotation.Autowired;
