+++ audit/src/main/java/gov/cms/ab2d/audit/remote/JobAuditClient.java
@@ -1,7 +1,7 @@
 package gov.cms.ab2d.audit.remote;
 
 import gov.cms.ab2d.common.dto.StaleJob;
-import gov.cms.ab2d.common.service.JobService;
+import gov.cms.ab2d.job.service.JobService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
