+++ job/src/main/java/gov/cms/ab2d/job/service/InvalidJobAccessException.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.common.service;
+package gov.cms.ab2d.job.service;
 
 public class InvalidJobAccessException extends RuntimeException {
