+++ job/src/main/java/gov/cms/ab2d/job/service/JobOutputServiceImpl.java
@@ -1,8 +1,9 @@
-package gov.cms.ab2d.common.service;
+package gov.cms.ab2d.job.service;
 
 import gov.cms.ab2d.common.model.Job;
 import gov.cms.ab2d.common.model.JobOutput;
 import gov.cms.ab2d.common.repository.JobOutputRepository;
+import gov.cms.ab2d.common.service.ResourceNotFoundException;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
