+++ job/src/test/resources/job-test.properties
@@ -1,0 +1,17 @@
+efs.mount=${java.io.tmpdir}/jobdownloads/
+
+audit.files.ttl.hours=72
+
+spring.datasource.url=${DB_URL}?rewriteBatchedInserts=true
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
+
+spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+
+spring.jpa.hibernate.ddl-auto=none
+spring.liquibase.enabled=true
+spring.liquibase.contexts=test
+
+logging.level.liquibase=ERROR
