+++ worker/pom.xml
@@ -22,6 +22,11 @@
         </dependency>
         <dependency>
             <groupId>gov.cms.ab2d</groupId>
+            <artifactId>job</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>gov.cms.ab2d</groupId>
             <artifactId>ab2d-bfd</artifactId>
             <version>1.2</version>
         </dependency>
