+++ worker/src/test/java/gov/cms/ab2d/worker/service/WorkerServiceStub.java
@@ -3,9 +3,9 @@ package gov.cms.ab2d.worker.service;
 import gov.cms.ab2d.common.model.Job;
 import gov.cms.ab2d.common.model.JobStatus;
 import gov.cms.ab2d.common.service.FeatureEngagement;
-import gov.cms.ab2d.common.service.JobService;
 import gov.cms.ab2d.common.service.PropertiesService;
 import gov.cms.ab2d.common.util.Constants;
+import gov.cms.ab2d.job.service.JobService;
 
 public class WorkerServiceStub implements WorkerService {
